+++ crate/operator/operations.py
@@ -54,6 +54,7 @@ from crate.operator.utils.kubeapi import (
 )
 from crate.operator.webhooks import (
     WebhookEvent,
+    WebhookOperation,
     WebhookStatus,
     WebhookTemporaryFailurePayload,
 )
@@ -159,6 +160,19 @@ async def get_pods_in_statefulset(
     return [{"uid": p.metadata.uid, "name": p.metadata.name} for p in all_pods.items]
 
 
+async def _send_update_progress_notification(
+    *, namespace: str, name: str, message: str, logger: logging.Logger
+):
+    await crate.send_feedback_notification(
+        namespace=namespace,
+        name=name,
+        message=message,
+        operation=WebhookOperation.UPDATE,
+        status=WebhookStatus.IN_PROGRESS,
+        logger=logger,
+    )
+
+
 async def restart_cluster(
     core: CoreV1Api,
     namespace: str,
@@ -206,6 +220,14 @@ async def restart_cluster(
     if next_pod_uid in all_pod_uids:
         # The next to-be-terminated pod still appears to be running.
         logger.info("Terminating pod '%s'", next_pod_name)
+        await _send_update_progress_notification(
+            namespace=namespace,
+            name=name,
+            message="Waiting for node "
+            f"{int(next_pod_name[next_pod_name.rindex('-')+1:])+1}/{len(all_pod_uids)}"
+            " to be terminated...",
+            logger=logger,
+        )
         # Trigger deletion of Pod.
         # This may take a while as it tries to gracefully stop the containers
         # of the Pod.
@@ -217,6 +239,14 @@ async def restart_cluster(
     elif next_pod_name in all_pod_names:
         total_nodes = get_total_nodes_count(old["spec"]["nodes"], "all")
         # The new pod has been spawned. Only a matter of time until it's ready.
+        await _send_update_progress_notification(
+            namespace=namespace,
+            name=name,
+            message="Waiting for node "
+            f"{int(next_pod_name[next_pod_name.rindex('-')+1:])+1}/{len(all_pod_uids)}"
+            " to be restarted...",
+            logger=logger,
+        )
         password, host = await asyncio.gather(
             get_system_user_password(core, namespace, name),
             get_host(core, namespace, name),